// frontend/src/types/api/listAutoRepliesApiAutoRepliesGetSort.ts
/**
 * Generated by orval v8.9.0 🍺
 * Do not edit manually.
 * mailassist
 * OpenAPI spec version: 0.1.0
 */

export type ListAutoRepliesApiAutoRepliesGetSort = typeof ListAutoRepliesApiAutoRepliesGetSort[keyof typeof ListAutoRepliesApiAutoRepliesGetSort];


export const ListAutoRepliesApiAutoRepliesGetSort = {
  newest: 'newest',
  oldest: 'oldest',
} as const;